style: UI改善と画面遷移をつなげた
